+++ lib/sourcescrub/models/concerns/company_items.rb
@@ -5,9 +5,10 @@ module Sourcescrub
   module Models
     # Tag
     class CompanyItems < Entity
-      attr_accessor :domain, :total, :items, :type
+      attr_accessor :identifier, :total, :items, :type
 
-      def parse_response_items(domain, kclass_name, response)
+      # the identifier can be domain or SS ID
+      def parse_response_items(identifier, kclass_name, response)
         headers = response.dig('headers')
         headers&.keys&.each do |attr_name|
           self.class.send(:define_method, attr_name.gsub('-', '_').to_sym) do
@@ -15,7 +16,7 @@ module Sourcescrub
           end
         end
 
-        dynamic_define_method(self, 'domain', domain)
+        dynamic_define_method(self, 'identifier', identifier)
         dynamic_define_method(self, 'type', kclass_name)
         dynamic_define_method(self, 'total', response.dig(total_key) || 0)
         dynamic_define_method(self, 'items', company_items(kclass_name, response.dig(items_key) || []))